Move functions.h and related definitions to legacy
