+++ conda/recipes/libcudf/meta.yaml
@@ -58,7 +58,7 @@ test:
     - test -f $PREFIX/include/cudf/cudf.h
     - test -f $PREFIX/include/cudf/datetime.hpp
     - test -f $PREFIX/include/cudf/filling.hpp
-    - test -f $PREFIX/include/cudf/functions.h
+    - test -f $PREFIX/include/cudf/legacy/functions.h
     - test -f $PREFIX/include/cudf/gis.hpp
     - test -f $PREFIX/include/cudf/groupby.hpp
     - test -f $PREFIX/include/cudf/io_functions.h
