+++ cpp/include/cudf/legacy/functions.h
