+++ cpp/src/hash/legacy/hashing.cu
@@ -23,7 +23,7 @@
 #include <table/legacy/device_table.cuh>
 #include "hash/hash_functions.cuh"
 #include <utilities/int_fastdiv.h>
-#include <utilities/nvtx/nvtx_utils.h>
+#include <utilities/nvtx/legacy/nvtx_utils.h>
 #include <copying/legacy/scatter.hpp>
 #include <cudf/types.hpp>
