+++ cpp/src/utilities/legacy/cuda_utils.cu
