+++ cpp/src/utilities/legacy/error_utils.cpp
