+++ cpp/tests/CMakeLists.txt
@@ -167,10 +167,10 @@ ConfigureTest(JOIN_TEST "${JOIN_TEST_SRC}")
 ###################################################################################################
 # - orderby tests ---------------------------------------------------------------------------------
 
-set(ORDERBY_TEST_SRC 
-    "${CMAKE_CURRENT_SOURCE_DIR}/orderby/orderby_tests.cu")
+set(LEGACY_ORDERBY_TEST_SRC 
+    "${CMAKE_CURRENT_SOURCE_DIR}/orderby/legacy/orderby_tests.cu")
 
-ConfigureTest(ORDERBY_TEST "${ORDERBY_TEST_SRC}")
+ConfigureTest(LEGACY_ORDERBY_TEST "${LEGACY_ORDERBY_TEST_SRC}")
 
 ###################################################################################################
 # - is_sorted tests ---------------------------------------------------------------------------------
@@ -208,19 +208,19 @@ ConfigureTest(LEGACY_BIT_MASK_TEST "${LEGACY_BIT_MASK_TEST_SRC}")
 ###################################################################################################
 # - datetime tests --------------------------------------------------------------------------------
 
-set(DATETIME_TEST_SRC
-    "${CMAKE_CURRENT_SOURCE_DIR}/datetime/datetime_ops_test.cu")
+set(LEGACY_DATETIME_TEST_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/datetime/legacy/datetime_ops_test.cu")
 
-ConfigureTest(DATETIME_TEST "${DATETIME_TEST_SRC}")
+ConfigureTest(LEGACY_DATETIME_TEST "${LEGACY_DATETIME_TEST_SRC}")
 
 ###################################################################################################
 # - hashing tests ---------------------------------------------------------------------------------
 
-set(HASHING_TEST_SRC
-    "${CMAKE_CURRENT_SOURCE_DIR}/hashing/hash_partition_test.cu"
-    "${CMAKE_CURRENT_SOURCE_DIR}/hashing/hash_test.cu")
+set(LEGACY_HASHING_TEST_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/hashing/legacy/hash_partition_test.cu"
+    "${CMAKE_CURRENT_SOURCE_DIR}/hashing/legacy/hash_test.cu")
 
-ConfigureTest(HASHING_TEST "${HASHING_TEST_SRC}")
+ConfigureTest(LEGACY_HASHING_TEST "${LEGACY_HASHING_TEST_SRC}")
 
 ###################################################################################################
 # - hash_map tests --------------------------------------------------------------------------------
@@ -331,10 +331,10 @@ ConfigureTest(ORC_TEST "${ORC_TEST_SRC}")
 ###################################################################################################
 # - sort tests ------------------------------------------------------------------------------------
 
-set(SORT_TEST_SRC
-    "${CMAKE_CURRENT_SOURCE_DIR}/sort/digitize_test.cu")
+set(LEGACY_SORT_TEST_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/sort/legacy/digitize_test.cu")
 
-ConfigureTest(SORT_TEST "${SORT_TEST_SRC}")
+ConfigureTest(LEGACY_SORT_TEST "${LEGACY_SORT_TEST_SRC}")
 
 ###################################################################################################
 # - types tests -----------------------------------------------------------------------------------
@@ -356,10 +356,10 @@ ConfigureTest(NVCATEGORY_TEST "${NVCATEGORY_TEST_SRC}")
 ###################################################################################################
 # - DLPack tests -------------------------------------------------------------------------------------
 
-set(DLPACK_TEST_SRC
-    "${CMAKE_CURRENT_SOURCE_DIR}/io/convert/dlpack_test.cu")
+set(LEGACY_DLPACK_TEST_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/io/convert/legacy/dlpack_test.cu")
 
-ConfigureTest(DLPACK_TEST "${DLPACK_TEST_SRC}")
+ConfigureTest(LEGACY_DLPACK_TEST "${LEGACY_DLPACK_TEST_SRC}")
 
 ###################################################################################################
 # - copying tests ---------------------------------------------------------------------------------
@@ -412,10 +412,10 @@ ConfigureTest(DEVICE_ATOMICS_TEST  "${DEVICE_ATOMICS_TEST_SRC}")
 ###################################################################################################
 # - transpose tests -------------------------------------------------------------------------------
 
-set(TRANSPOSE_TEST_SRC
-    "${CMAKE_CURRENT_SOURCE_DIR}/transpose/transpose_test.cu")
+set(LEGACY_TRANSPOSE_TEST_SRC
+    "${CMAKE_CURRENT_SOURCE_DIR}/transpose/legacy/transpose_test.cu")
 
-ConfigureTest(TRANSPOSE_TEST "${TRANSPOSE_TEST_SRC}")
+ConfigureTest(LEGACY_TRANSPOSE_TEST "${LEGACY_TRANSPOSE_TEST_SRC}")
 
 ###################################################################################################
 # - legacy table tests -----------------------------------------------------------------------------------
