+++ cpp/tests/hashing/legacy/hash_test.cu
